+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Changed: Bump dependency versions to the latest version at 2024-09
+- Changed: create all Front Matter before creating pages
 
 ## v[4.3.0] 2024-09-05